E2.Term

Documentation

>>> :set -XOverloadedStrings

newtype Variable

A variable token, identified by its name.

Constructors

Variable

Fields

getVariable :: String

newtype Label

A label (such as constructor label).

Constructors

Label

Fields

getLabel :: String

data Term metavar var

A term, parametrised by the type of its metavariables and variables.

Constructors

Var var
A variable term.

Con Label [ArgTerm metavar var]
A constructor (a.k.a. function symbol), fully applied to all its arguments.

MetaVar metavar [Term metavar var]
A parametrised metavariable.

data ArgTerm metavar var

An argument term (i.e. occuring in the list of arguments to a constructor).

Constructors

Regular (Term metavar var)
Regular subterm.

Scoped (Term metavar (Inc var))
Scoped subterm (introduces one bound variable).

type Term' = Term Variable Variable

A convenient version of Term.

type ArgTerm' = ArgTerm Variable

metavars :: Term m a -> [m]

Extract a list of all metavariable names that occur in a given term.

metavarsArg :: ArgTerm m a -> [m]

Extract a list of all metavariable names that occur in a given argment term.

Substitutions

Variable substitution

substitute :: (a -> Term m b) -> Term m a -> Term m b

Substitute all variables in a term.

substituteArg :: (a -> Term m b) -> ArgTerm m a -> ArgTerm m b

Substitute all variables in an argument term.

Meta substitution

substituteMeta :: (m -> Term n (IncMany a)) -> Term m a -> Term n a

Substitute all meta variables in a term. Note that here all metavariables must have a substitution defined.

substituteMetaArg :: (m -> Term n (IncMany a)) -> ArgTerm m a -> ArgTerm n a

Substitute all meta variables in an argument term. Note that here all metavariables must have a substitution defined.

substituteMeta' :: (m -> Maybe (Term m (IncMany a))) -> Term m a -> Term m a

Substitute some meta variables in a term. Here, Nothing means that corresponding meta variable should not be substituted (but substitution may still take place in its parameters).

substituteMetaArg' :: (m -> Maybe (Term m (IncMany a))) -> ArgTerm m a -> ArgTerm m a

Substitute some meta variables in an argument term. Here, Nothing means that corresponding meta variable should not be substituted (but substitution may still take place in its parameters).

Renaming meta variables

renameMeta :: (m -> n) -> Term m a -> Term n a

Rename meta variables in a term.

renameMetaArg :: (m -> n) -> ArgTerm m a -> ArgTerm n a

Rename meta variables in an argument term.

Default fresh names

defaultFreshVars :: IsString a => [a]

Default fresh names for variables:

>>> mapM_ putStrLn $ take 5 defaultFreshVars
x₁
x₂
x₃
x₄
x₅

defaultFreshBoundVars :: IsString a => [a]

Default stream of fresh names for bound variables.

>>> mapM_ putStrLn $ take 5 defaultFreshBoundVars
z₁
z₂
z₃
z₄
z₅

defaultFreshMetaVars :: IsString a => [a]

Default strem of fresh names for meta variables.

>>> mapM_ putStrLn $ take 5 defaultFreshMetaVars
m₁
m₂
m₃
m₄
m₅

mkDefaultFreshVars :: IsString a => String -> [a]

Make an infinite stream of default names with numeric subscript.

>>> mapM_ putStrLn $ take 5 (mkDefaultFreshVars "example")
example₁
example₂
example₃
example₄
example₅

Pretty-printing

ppTerm' :: Term' -> String

Pretty print a Term'.

>>> t = Con "APP" [Regular (Con "LAM" [Scoped (MetaVar "m1" [Var Z])]), Regular (MetaVar "m2" [])]
>>> putStrLn $ ppTerm' t
APP(LAM(x₁.m1[x₁]), m2[])

ppTerm :: [String] -> (a -> String) -> Term Variable a -> String

Pretty print a Term.

ppArgTerm

Arguments

:: [String]
A (possibly infinite) list of fresh names for bound variables.

-> (a -> String)
A pretty printing function for free variables.

-> ArgTerm Variable a
An argument term.

-> String

Pretty print an argument term (regular or scoped).

ppInc :: String -> Inc String -> String

Pretty-print a variable (using a default name for the bound variable).

ppIncMany :: [String] -> IncMany String -> String

Pretty-print a variable (using a list of default names for the bound variables).

Helpers

(!!?) :: [a] -> String -> Int -> a

A version of !! with a custom error message.

De Bruijn indices as nested data types

data Inc var

Add one (bound) variable.

Constructors

Z
The bound variable.

S var
A free variable (or another bound variable from the outside).

data IncMany var

Add arbitrarily many bound variables.

Constructors

B Int
The bound variable.

F var
A free variable (or another bound variable from the outside).

instantiateMany :: Monad f => (Int -> f a) -> f (IncMany a) -> f a

Instantiate bound variables in a generalized term.

joinInc :: Inc (IncMany a) -> IncMany a

Convert two scopes into one.

commuteInc :: Inc (IncMany a) -> IncMany (Inc a)

Commute scopes.

removeInc :: Alternative f => Inc a -> f a

Remove bound variables if possible.

removeIncMany :: Alternative f => IncMany a -> f a

Remove bound variables if possible.

decIncMany :: Alternative f => IncMany a -> f (IncMany a)

Remove the first bound variable if possible.

removeAllIncMany :: (Traversable t, Alternative f) => t (IncMany a) -> f (t a)

Remove the first bound variable from a generalised term.