E2.Term.Unification

Documentation

data Forall t var

Quantify over arbitrarily many variables using Inc.

Constructors

Forall (Forall t (Inc var))
\(\forall x. C\)

Ground (t var)
Ground term/constraint (no quantifiers).

data Rule metavar a

A rule \(t \to u\).

Constructors

(Term metavar a) :-> (Term metavar a)

ruleLHS :: Rule m Void -> Term m Void

Extract rule's left hand side.

rulesLHS :: [Rule m Void] -> [Term m Void]

Extract left hand side from a collection of rules.

data GroundConstraint metavar var

A ground constraint has no quantifiers.

Constructors

(Term metavar var) :==: (Term metavar var)
\( t \stackrel{?}{=} u\)

type Constraint metavar = Forall (GroundConstraint metavar)

A constraint is a ground constraint put under some number of quantifiers.

\(\forall \overline{x_k}. t \stackrel{?}{=} u\)

data MetaSubst m var

A meta variable substitution.

Constructors

MetaSubst

Fields

metaSubstVar :: m

Meta variable to be substituted.

metaSubstArity :: Int

Meta variable arity (for checks).

metaSubstBody :: Term m (IncMany var)

Meta variable body (a scoped term).

metaSubst :: m -> Int -> Term m (IncMany var) -> MetaSubst m var

Construct a meta variable substitution and check that its declared arity is not exceeded in the body.

newtype MetaSubsts m var

A collection of simultaneous meta variable substitutions.

Constructors

MetaSubsts

Fields

getMetaSubsts :: [MetaSubst m var]

toLookupList :: MetaSubsts m var -> [(m, Term m (IncMany var))]

Convert meta substitutions into a lookup list.

applySubsts :: Eq m => MetaSubsts m a -> Constraint m a -> Constraint m a

Apply meta substitutions to a given constraint.

newtype Unify v a

The unification context (a combination of backtracking and fresh name supply).

Constructors

Unify

Fields

runUnify :: LogicT (Fresh v) a

defaultUnify :: Unify Variable a -> a

Run a unification computation using default fresh name supply and extracting the first result.

NOTE: will throw an exception when no solution is available.

\(E^2\)-unification procedure

tryE2

Arguments

:: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f, MonadLogic f)

=> [Rule m Void]
Set \(R\) of rewrite rules.

-> [Constraint m a]
Collection of constraints to simplify.

-> f ([Constraint m a], MetaSubsts m a)

Attempt applying one step of \(E^2\)-unification to a collection of constraints.

This will try to simplify constraints in the following way:

If there are no constraints to simplify, then exit successfully.

Apply delete or eliminate rule to some constraint if possible.

Simplify any rigid-rigid constraint with decompose or mutate rules.

Otherwise, apply project, imitate, or mutateMeta rules.

unifyDFS

Arguments

:: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f, MonadLogic f)

=> Int
Maximum search depth.

-> [Rule m Void]
Set \(R\) of rewrite rules.

-> [Constraint m a]
Collection of constraints.

-> f ([Constraint m a], MetaSubsts m a)

Perform unificiation procedure, using DFS and limiting search to a fixed maximum depth. If maximum depth is reached, but not all constraints are simplified, leftover constraints are returned along with intermediate substitution.

unifyBFSviaDFS

Arguments

:: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f, MonadLogic f)

=> Int
Number of DFS interations.

-> Int
Max depth per one DFS iteration.

-> [Rule m Void]
Set \(R\) of rewrite rules.

-> [Constraint m a]
Collection of constraints.

-> f (MetaSubsts m a)

Perform unificiation procedure, using k-depth DFS and limiting search to a fixed maximum depth and BFS iterations. If maximum depth is reached, but not all constraints are simplified, these solutions are discarded.

tryClose :: MonadLogic f => f ([Constraint m a], MetaSubsts m a) -> f (Either [([Constraint m a], MetaSubsts m a)] (MetaSubsts m a))

Attempt to find a solution without any leftover constraints. If such a solution exists, return corresponding substitution (using Right). Otherwise, collect all partial solutions in a list (using Left).

constraintMetaVars :: Constraint m a -> [m]

Extract all meta variables from a constraint.

isRigidRigid :: Constraint m a -> Bool

Check if constraint is rigid-rigid (i.e. both sides are not meta variables).

Pretty-printing

ppConstraints' :: [Constraint Variable Variable] -> String

ppConstraint' :: Constraint Variable Variable -> String

ppConstraint :: [String] -> (a -> String) -> Constraint Variable a -> String

ppMetaSubsts' :: MetaSubsts Variable Variable -> String

ppMetaSubsts :: [String] -> (a -> String) -> MetaSubsts Variable a -> String

ppMetaSubst :: [String] -> (a -> String) -> MetaSubst Variable a -> String

ppRules :: [Rule Variable Void] -> String

\(E^2\)-unification rules

delete :: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f) => GroundConstraint m (IncMany a) -> f ([Constraint m (IncMany a)], MetaSubsts m (IncMany a))

Delete rule: \(\forall \overline{x_k}. t \stackrel{?}{=} t \longrightarrow \langle \varnothing, \mathsf{id} \rangle \)

decompose :: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f) => GroundConstraint m (IncMany a) -> f ([Constraint m (IncMany a)], MetaSubsts m (IncMany a))

Decompose rule: \(\forall \overline{x_k}. F(\overline{t_n}, \overline{x.s_m}) \stackrel{?}{=} F(\overline{t'_n}, \overline{x.s'_m}) \longrightarrow \langle C_1 \cup C_2, \mathsf{id} \rangle \) where

\( C_1 = \{ \forall \overline{x_k}. t_i \stackrel{?}{=} t'_i \) for all \(i = 1 \ldots n \}\)

\( C_2 = \{ \forall \overline{x_k}, x. s_j \stackrel{?}{=} s'_j \) for all \(j = 1 \ldots m \}\)

eliminate :: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f) => GroundConstraint m (IncMany a) -> f ([Constraint m (IncMany a)], MetaSubsts m (IncMany a))

Eliminate rule: \(\forall \overline{x_k}. m[\overline{t_n}] \stackrel{?}{=} u \longrightarrow \langle \varnothing, \sigma \rangle \) where

\(\sigma = [m[\overline{z_n} \mapsto u[\overline{t_n} \mapsto \overline{z_k}]]]\)

\(\overline{t_k}\) is a list of distinct bound variables (from \(\overline{x_k}\))

\(m\) does not occur in \(u\)

imitate :: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f) => GroundConstraint m (IncMany a) -> f ([Constraint m (IncMany a)], MetaSubsts m (IncMany a))

Imitate rule: \(\forall \overline{x_k}. m[\overline{t_k}] \stackrel{?}{=} F(\overline{u_p}, \overline{x.s_q}) \longrightarrow \langle C_1 \cup C_2, \sigma \rangle \) where

\(\sigma = [m[\overline{z_n}] \mapsto F(\overline{m_p[\overline{z_k}]}, \overline{x.m'_q[x, \overline{z_k}]})] \)

\( C_1 = \{ \forall \overline{x_k}. m_i[\overline{t_k}] \stackrel{?}{=} u_i \) for all \(i = 1 \ldots p \}\)

\( C_2 = \{ \forall \overline{x_k}, x. m'_j[x, \overline{t_k}] \stackrel{?}{=} s'_j \) for all \(j = 1 \ldots q \}\)

\(\overline{m_p}, \overline{m'_q}\) are fresh meta variables

\(m\) does not occur in \(\overline{u_p}\) or \(\overline{s_q}\)

project :: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f) => GroundConstraint m (IncMany a) -> f ([Constraint m (IncMany a)], MetaSubsts m (IncMany a))

Project rule: \(\forall \overline{x_k}. m[\overline{t_k}] \stackrel{?}{=} u \longrightarrow \langle C, \sigma \rangle \) where

\(\sigma = [m[\overline{z_n}] \mapsto z_i] \)

\( C = \{ \forall \overline{x_k}. t_i \stackrel{?}{=} u \}\)

\(m\) does not occur in \(u\)

mutate :: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f) => [Rule m Void] -> [Term m (IncMany a)] -> GroundConstraint m (IncMany a) -> f ([Constraint m (IncMany a)], MetaSubsts m (IncMany a))

Mutate rule: \(\forall \overline{x_k}. F(\overline{t_p}, \overline{x.s_q}) \stackrel{?}{=} u \longrightarrow \langle C_1 \cup C_2 \cup C_3, \mathsf{id} \rangle \) where

\(F(\overline{l_p}, \overline{x.l'_q}) \to r\) is in \(R\)

\( C_1 = \{ \forall \overline{x_k}. t_i \stackrel{?}{=} l_i \) for all \(i = 1 \ldots p \}\)

\( C_2 = \{ \forall \overline{x_k}, x. s_j \stackrel{?}{=} l'_j \) for all \(j = 1 \ldots m \}\)

\( C_3 = \{ \forall \overline{x_k}. r \stackrel{?}{=} u \} \)

mutateMeta :: (Eq m, Eq a, MonadPlus f, MonadFail f, MonadFresh m f) => [Rule m Void] -> [Term m (IncMany a)] -> GroundConstraint m (IncMany a) -> f ([Constraint m (IncMany a)], MetaSubsts m (IncMany a))

Mutate (meta) rule: \(\forall \overline{x_k}. m[\overline{t_k}] \stackrel{?}{=} u \longrightarrow \langle C_1 \cup C_2 \cup C_3, \mathsf{\sigma} \rangle \) where

\(F(\overline{l_p}, \overline{x.l'_q}) \to r\) is in \(R\)

\(\sigma = [m[\overline{z_n}] \mapsto F(\overline{m_p[\overline{z_k}]}, \overline{x.m'_q[x, \overline{z_k}]})] \)

\( C_1 = \{ \forall \overline{x_k}. m_i[\overline{t_k}] \stackrel{?}{=} l_i \) for all \(i = 1 \ldots p \}\)

\( C_2 = \{ \forall \overline{x_k}, x. m'_j[x, \overline{t_k}] \stackrel{?}{=} l'_j \) for all \(j = 1 \ldots m \}\)

\( C_3 = \{ \forall \overline{x_k}. r \stackrel{?}{=} u \} \)

\(m\) does not occur in \(u\)

Helpers

distinctBoundVars :: [Term m (IncMany a)] -> Maybe (IncMany a -> Term m (IncMany (IncMany a)))

Check whether given list of terms is a list of distinct bound variables. Return the function appropriate to use in substitution.

argTermToMetaVar

Arguments

:: MonadFresh m f

=> [Term m var]
Parameters of the meta variable on the LHS of the imitate/mutateMeta rule.

-> ArgTerm m var
Argument term of a constructor.

-> WriterT [Constraint m var] f (ArgTerm m (IncMany var))

Generate a fresh meta variable corresponding to an argument of a constructor.

Used in imitate and mutateMeta rules.

freshRule :: (Eq m, MonadFresh m f) => [Term m a] -> Rule m Void -> f (Rule m a)

Generate a fresh version of a rewrite rule given a list of additional parameters for the meta variables.

Used in mutate and mutateMeta rules.

Example TRS

Lambda calculus with pairs

rulesLambda :: [Rule Variable Void]

Rewrite rules for lambda calculus with pairs.

app :: Term Variable a -> Term Variable a -> Term Variable a

A helper to create application terms \(t_1\;t_2\)

lam :: Term Variable (Inc a) -> Term Variable a

A helper to create lambda abstraction terms \(\lambda x. t\)

first :: Term Variable a -> Term Variable a

A helper to create first project terms \(\pi_1\;t\)

second :: Term Variable a -> Term Variable a

A helper to create second project terms \(\pi_2\;t\)

pair :: Term Variable a -> Term Variable a -> Term Variable a

A helper to create pair terms \(\langle t_1, t_2 \rangle\)

Example terms and constraints

ex1 :: Term Variable a

>>> putStrLn $ ppTerm' ex1
APP(LAM(x₁.m1[x₁]), m2[])

ex2 :: Term Variable a

>>> putStrLn $ ppTerm' ex2
m1[m2[]]

ex3 :: Term'

>>> putStrLn $ ppTerm' ex3
APP(m[], x)

ex4 :: Constraint Variable Variable

>>> putStrLn $ ppConstraint' ex4
APP(m[y], x) =?= y

ex5 :: Constraint Variable Variable

>>> putStrLn $ ppConstraint' ex5
forall x₁.forall x₂.m1[x₂, x₁] =?= APP(x₁, x₂)

>>> putStrLn $ ppMetaSubsts' $ defaultUnify $ unifyBFSviaDFS 1 5 rulesLambda [ ex5 ]
m1[x₁, x₂] -> APP(x₂, x₁)

ex6 :: Constraint Variable Variable

>>> putStrLn $ ppConstraint' ex6
APP(m[], PAIR(LAM(x₁.APP(x₁, y)), g)) =?= APP(g, y)

FIXME: unification does not finish in reasonable time!

ex7 :: Constraint Variable Variable

>>> putStrLn $ ppConstraint' ex7
m[y, g] =?= APP(g, y)
>>> putStrLn $ ppMetaSubsts' $ defaultUnify $ unifyBFSviaDFS 1 5 rulesLambda [ ex7 ]
m[x₁, x₂] -> APP(x₂, x₁)

ex8 :: Constraint Variable Variable

>>> putStrLn $ ppConstraint' ex8
APP(m[y], g) =?= APP(g, y)

FIXME: takes too long >>> putStrLn $ ppMetaSubsts' $ defaultUnify $ unifyBFSviaDFS 1 4 rulesLambda [ ex8 ] m[x₁] -> LAM(x₂.APP(x₂, x₁))

ex9 :: Constraint Variable Variable

>>> putStrLn $ ppConstraint' ex9
m[PAIR(y, g)] =?= g

FIXME: takes a bit too long and contains unresolved metavariable! >>> putStrLn $ ppMetaSubsts' $ defaultUnify $ unifyBFSviaDFS 1 4 rulesLambda [ ex9 ] m[x₁] -> APP(LAM(x₂.SECOND(x₁)), m₉[x₁])

>>> putStrLn $ ppMetaSubsts' $ defaultUnify $ unifyBFSviaDFS 1 5 rulesLambda [ ex9 ]
m[x₁] -> SECOND(x₁)

Fresh name supply monad

newtype FreshT v m a

A monad transformer with fresh name supply.

Constructors

FreshT

Fields

runFreshT :: StateT [v] m a

evalFreshT :: Monad m => FreshT v m a -> [v] -> m a

Evaluate a computation with given supply of fresh names.

type Fresh v = FreshT v Identity

Simple fresh name supply context.

evalFresh :: Fresh v a -> [v] -> a

Evaluate a computation with given supply of fresh names.

class Monad m => MonadFresh v m | m -> v where

A monad that is capable of generating fresh names of type v.

Methods

fresh :: m v

Generate a fresh name.

Helpers

selectOne :: Alternative f => [a] -> f (a, [a])

Non-deterministically select one element out of a list.

>>> selectOne [1..3] :: [(Int, [Int])]
[(1,[2,3]),(2,[1,3]),(3,[1,2])]

splits :: [a] -> [(a, [a])]

All possible splits of a list into an element and a list of other elements.

>>> splits [1..3]
[(1,[2,3]),(2,[1,3]),(3,[1,2])]

choose :: Alternative f => [a] -> f a

Choose a value from a list.

orElse :: MonadLogic m => m a -> m a -> m a

Try the second computation only if the first one fails (i.e. no backtracking).

The following examples show the difference between orElse and mplus for Logic:

>>> observeAll (pure 1 `orElse` pure 2)
[1]
>>> observeAll (pure 1 `mplus` pure 2)
[1,2]

>>> observeAll $ do { x <- pure 1 `orElse` pure 2; guard (even x); return x }
[]
>>> observeAll $ do { x <- pure 1 `mplus` pure 2; guard (even x); return x }
[2]

notrace :: String -> a -> a